pyproject.toml
[project]
name = "creative-writer"
version = "0.1.0"
description = "Self-improving LinkedIn content engine for PMM leaders"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.0",
    "python-docx>=1.1.0",
    "chromadb>=0.5.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
ui = [
    "streamlit>=1.30.0",
]

[project.scripts]
creative-writer = "main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
